requirements.txt
# flask\[async\]
requests
python-dotenv
flask[async]
Flask-CORS